requirements.txt
SpeechRecognition
pipwin
pyaudio
gtts
google-generativeai
python-dotenv
streamlit